增加对切割保存切割数据到数据库

+++ easytnt-imports-paper/src/main/java/com/easytnt/importpaper/bean/CountContainer.java
@@ -14,7 +14,7 @@ import java.util.concurrent.atomic.AtomicLong;
  */
 public class CountContainer<T> {
 	private CopyOnWriteArrayList<T> container = new CopyOnWriteArrayList<>();
-	private long totalNum = 0;
+	private int totalNum = 0;
 	private AtomicLong fileNumber = new AtomicLong();
 	private final int capacity;
 	private AtomicBoolean isOver = new AtomicBoolean(false);
@@ -50,11 +50,11 @@ public class CountContainer<T> {
 		this.isOver.set(isOver);
 	}
 
-	public long getTotalNum() {
+	public int getTotalNum() {
 		return totalNum;
 	}
 
-	public void setTotalNum(long totalNum) {
+	public void setTotalNum(int totalNum) {
 		this.totalNum = totalNum;
 	}
